+++ examples/data-sources/iosxr_mpls_ldp_address_family/data-source.tf
@@ -1,3 +1,0 @@
-data "iosxr_mpls_ldp_address_family" "example" {
-  af_name = "ipv4"
-}