+++ examples/data-sources/iosxr_mpls_ldp_interface/data-source.tf
@@ -1,3 +1,0 @@
-data "iosxr_mpls_ldp_interface" "example" {
-  interface_name = "GigabitEthernet0/0/0/1"
-}