+++ examples/resources/iosxr_mpls_ldp_address_family/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_mpls_ldp_address_family.example "Cisco-IOS-XR-um-mpls-ldp-cfg:/mpls/ldp/address-families/address-family[af-name=ipv4]"