+++ examples/resources/iosxr_mpls_ldp_address_family/resource.tf
@@ -1,3 +1,0 @@
-resource "iosxr_mpls_ldp_address_family" "example" {
-  af_name = "ipv4"
-}