+++ examples/resources/iosxr_mpls_ldp_interface/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_mpls_ldp_interface.example "Cisco-IOS-XR-um-mpls-ldp-cfg:/mpls/ldp/interfaces/interface[interface-name=GigabitEthernet0/0/0/1]"