+++ examples/resources/iosxr_mpls_ldp_interface/resource.tf
@@ -1,3 +1,0 @@
-resource "iosxr_mpls_ldp_interface" "example" {
-  interface_name = "GigabitEthernet0/0/0/1"
-}