+++ internal/provider/data_source_iosxr_mpls_ldp_test.go
@@ -17,6 +17,8 @@ func TestAccDataSourceIosxrMPLSLDP(t *testing.T) {
 				Config: testAccDataSourceIosxrMPLSLDPConfig,
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("data.iosxr_mpls_ldp.test", "router_id", "1.2.3.4"),
+					resource.TestCheckResourceAttr("data.iosxr_mpls_ldp.test", "address_families.0.af_name", "ipv4"),
+					resource.TestCheckResourceAttr("data.iosxr_mpls_ldp.test", "interfaces.0.interface_name", "GigabitEthernet0/0/0/1"),
 				),
 			},
 		},
@@ -27,6 +29,12 @@ const testAccDataSourceIosxrMPLSLDPConfig = `
 
 resource "iosxr_mpls_ldp" "test" {
 	router_id = "1.2.3.4"
+	address_families = [{
+		af_name = "ipv4"
+	}]
+	interfaces = [{
+		interface_name = "GigabitEthernet0/0/0/1"
+	}]
 }
 
 data "iosxr_mpls_ldp" "test" {