+++ internal/provider/model_iosxr_mpls_ldp.go
@@ -3,15 +3,27 @@
 package provider
 
 import (
+	"fmt"
+	"reflect"
+	"strconv"
+
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/tidwall/gjson"
 	"github.com/tidwall/sjson"
 )
 
 type MPLSLDP struct {
-	Device   types.String `tfsdk:"device"`
-	Id       types.String `tfsdk:"id"`
-	RouterId types.String `tfsdk:"router_id"`
+	Device          types.String             `tfsdk:"device"`
+	Id              types.String             `tfsdk:"id"`
+	RouterId        types.String             `tfsdk:"router_id"`
+	AddressFamilies []MPLSLDPAddressFamilies `tfsdk:"address_families"`
+	Interfaces      []MPLSLDPInterfaces      `tfsdk:"interfaces"`
+}
+type MPLSLDPAddressFamilies struct {
+	AfName types.String `tfsdk:"af_name"`
+}
+type MPLSLDPInterfaces struct {
+	InterfaceName types.String `tfsdk:"interface_name"`
 }
 
 func (data MPLSLDP) getPath() string {
@@ -23,6 +35,22 @@ func (data MPLSLDP) toBody() string {
 	if !data.RouterId.Null && !data.RouterId.Unknown {
 		body, _ = sjson.Set(body, "router-id", data.RouterId.Value)
 	}
+	if len(data.AddressFamilies) > 0 {
+		body, _ = sjson.Set(body, "address-families.address-family", []interface{}{})
+		for index, item := range data.AddressFamilies {
+			if !item.AfName.Null && !item.AfName.Unknown {
+				body, _ = sjson.Set(body, "address-families.address-family"+"."+strconv.Itoa(index)+"."+"af-name", item.AfName.Value)
+			}
+		}
+	}
+	if len(data.Interfaces) > 0 {
+		body, _ = sjson.Set(body, "interfaces.interface", []interface{}{})
+		for index, item := range data.Interfaces {
+			if !item.InterfaceName.Null && !item.InterfaceName.Unknown {
+				body, _ = sjson.Set(body, "interfaces.interface"+"."+strconv.Itoa(index)+"."+"interface-name", item.InterfaceName.Value)
+			}
+		}
+	}
 	return body
 }
 
@@ -32,12 +60,94 @@ func (data *MPLSLDP) updateFromBody(res []byte) {
 	} else {
 		data.RouterId.Null = true
 	}
+	for i := range data.AddressFamilies {
+		keys := [...]string{"af-name"}
+		keyValues := [...]string{data.AddressFamilies[i].AfName.Value}
+
+		var r gjson.Result
+		gjson.GetBytes(res, "address-families.address-family").ForEach(
+			func(_, v gjson.Result) bool {
+				found := false
+				for ik := range keys {
+					if v.Get(keys[ik]).String() == keyValues[ik] {
+						found = true
+						continue
+					}
+					found = false
+					break
+				}
+				if found {
+					r = v
+					return false
+				}
+				return true
+			},
+		)
+		if value := r.Get("af-name"); value.Exists() {
+			data.AddressFamilies[i].AfName.Value = value.String()
+		} else {
+			data.AddressFamilies[i].AfName.Null = true
+		}
+	}
+	for i := range data.Interfaces {
+		keys := [...]string{"interface-name"}
+		keyValues := [...]string{data.Interfaces[i].InterfaceName.Value}
+
+		var r gjson.Result
+		gjson.GetBytes(res, "interfaces.interface").ForEach(
+			func(_, v gjson.Result) bool {
+				found := false
+				for ik := range keys {
+					if v.Get(keys[ik]).String() == keyValues[ik] {
+						found = true
+						continue
+					}
+					found = false
+					break
+				}
+				if found {
+					r = v
+					return false
+				}
+				return true
+			},
+		)
+		if value := r.Get("interface-name"); value.Exists() {
+			data.Interfaces[i].InterfaceName.Value = value.String()
+		} else {
+			data.Interfaces[i].InterfaceName.Null = true
+		}
+	}
 }
 
 func (data *MPLSLDP) fromBody(res []byte) {
 	if value := gjson.GetBytes(res, "router-id"); value.Exists() {
 		data.RouterId.Value = value.String()
 		data.RouterId.Null = false
+	}
+	if value := gjson.GetBytes(res, "address-families.address-family"); value.Exists() {
+		data.AddressFamilies = make([]MPLSLDPAddressFamilies, 0)
+		value.ForEach(func(k, v gjson.Result) bool {
+			item := MPLSLDPAddressFamilies{}
+			if cValue := v.Get("af-name"); cValue.Exists() {
+				item.AfName.Value = cValue.String()
+				item.AfName.Null = false
+			}
+			data.AddressFamilies = append(data.AddressFamilies, item)
+			return true
+		})
+	}
+	if value := gjson.GetBytes(res, "interfaces.interface"); value.Exists() {
+		data.Interfaces = make([]MPLSLDPInterfaces, 0)
+		value.ForEach(func(k, v gjson.Result) bool {
+			item := MPLSLDPInterfaces{}
+			if cValue := v.Get("interface-name"); cValue.Exists() {
+				item.InterfaceName.Value = cValue.String()
+				item.InterfaceName.Null = false
+			}
+			data.Interfaces = append(data.Interfaces, item)
+			return true
+		})
 	}
 }
 
@@ -58,14 +168,87 @@ func (data *MPLSLDP) setUnknownValues() {
 		data.RouterId.Unknown = false
 		data.RouterId.Null = true
 	}
+	for i := range data.AddressFamilies {
+		if data.AddressFamilies[i].AfName.Unknown {
+			data.AddressFamilies[i].AfName.Unknown = false
+			data.AddressFamilies[i].AfName.Null = true
+		}
+	}
+	for i := range data.Interfaces {
+		if data.Interfaces[i].InterfaceName.Unknown {
+			data.Interfaces[i].InterfaceName.Unknown = false
+			data.Interfaces[i].InterfaceName.Null = true
+		}
+	}
 }
 
 func (data *MPLSLDP) getDeletedListItems(state MPLSLDP) []string {
 	deletedListItems := make([]string, 0)
+	for i := range state.AddressFamilies {
+		keys := [...]string{"af-name"}
+		stateKeyValues := [...]string{state.AddressFamilies[i].AfName.Value}
+
+		emptyKeys := true
+		if !reflect.ValueOf(state.AddressFamilies[i].AfName.Value).IsZero() {
+			emptyKeys = false
+		}
+		if emptyKeys {
+			continue
+		}
+
+		found := false
+		for j := range data.AddressFamilies {
+			found = true
+			if state.AddressFamilies[i].AfName.Value != data.AddressFamilies[j].AfName.Value {
+				found = false
+			}
+			if found {
+				break
+			}
+		}
+		if !found {
+			keyString := ""
+			for ki := range keys {
+				keyString += "[" + keys[ki] + "=" + stateKeyValues[ki] + "]"
+			}
+			deletedListItems = append(deletedListItems, fmt.Sprintf("%v/address-families/address-family%v", state.getPath(), keyString))
+		}
+	}
+	for i := range state.Interfaces {
+		keys := [...]string{"interface-name"}
+		stateKeyValues := [...]string{state.Interfaces[i].InterfaceName.Value}
+
+		emptyKeys := true
+		if !reflect.ValueOf(state.Interfaces[i].InterfaceName.Value).IsZero() {
+			emptyKeys = false
+		}
+		if emptyKeys {
+			continue
+		}
+
+		found := false
+		for j := range data.Interfaces {
+			found = true
+			if state.Interfaces[i].InterfaceName.Value != data.Interfaces[j].InterfaceName.Value {
+				found = false
+			}
+			if found {
+				break
+			}
+		}
+		if !found {
+			keyString := ""
+			for ki := range keys {
+				keyString += "[" + keys[ki] + "=" + stateKeyValues[ki] + "]"
+			}
+			deletedListItems = append(deletedListItems, fmt.Sprintf("%v/interfaces/interface%v", state.getPath(), keyString))
+		}
+	}
 	return deletedListItems
 }
 
 func (data *MPLSLDP) getEmptyLeafsDelete() []string {
 	emptyLeafsDelete := make([]string, 0)
+
 	return emptyLeafsDelete
 }